Move Plotly settle helpers into their own module
