+++ shinywidgets/static/output.js
@@ -36,7 +36,17 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \***********************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @jupyter-widgets/html-manager */ \"@jupyter-widgets/html-manager\");\n/* harmony import */ var _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _comm__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./comm */ \"./src/comm.ts\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./utils */ \"./src/utils.ts\");\nvar _a;\n\n\n\n/******************************************************************************\n * Define a custom HTMLManager for use with Shiny\n ******************************************************************************/\nclass OutputManager extends _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__.HTMLManager {\n    // In a soon-to-be-released version of @jupyter-widgets/html-manager,\n    // display_view()'s first \"dummy\" argument will be removed... this shim simply\n    // makes it so that our manager can work with either version\n    // https://github.com/jupyter-widgets/ipywidgets/commit/159bbe4#diff-45c126b24c3c43d2cee5313364805c025e911c4721d45ff8a68356a215bfb6c8R42-R43\n    async display_view(view, options) {\n        const n_args = super.display_view.length;\n        if (n_args === 3) {\n            return super.display_view({}, view, options);\n        }\n        else {\n            // @ts-ignore\n            return super.display_view(view, options);\n        }\n    }\n}\n// Define our own custom module loader for Shiny\nconst shinyRequireLoader = async function (moduleName, moduleVersion) {\n    // shiny provides a shim of require.js which allows <script>s with anonymous\n    // define()s to be loaded without error. When an anonymous define() occurs,\n    // the shim uses the data-requiremodule attribute (set by require.js) on the script\n    // to determine the module name.\n    // https://github.com/posit-dev/py-shiny/blob/230940c/scripts/define-shims.js#L10-L16\n    // In the context of shinywidgets, when a widget gets rendered, it should\n    // come with another <script> tag that does `require.config({paths: {...}})`\n    // which maps the module name to a URL of the widget's JS file.\n    const oldAmd = window.define.amd;\n    // This is probably not necessary, but just in case -- especially now in a\n    // anywidget/ES6 world, we probably don't want to load AMD modules\n    // (plotly is one example of a widget that will fail to load if AMD is enabled)\n    window.define.amd = false;\n    // Store jQuery global since loading we load a module, it may overwrite it\n    // (qgrid is one good example)\n    const old$ = window.$;\n    const oldJQ = window.jQuery;\n    if (moduleName === 'qgrid') {\n        // qgrid wants to use base/js/dialog (if it's available) for full-screen tables\n        // https://github.com/quantopian/qgrid/blob/877b420/js/src/qgrid.widget.js#L11-L16\n        // Maybe that's worth supporting someday, but for now, we define it to be nothing\n        // to avoid require('qgrid') from producing an error\n        window.define(\"base/js/dialog\", [], function () { return null; });\n    }\n    return (0,_jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__.requireLoader)(moduleName, moduleVersion).finally(() => {\n        window.define.amd = oldAmd;\n        window.$ = old$;\n        window.jQuery = oldJQ;\n    });\n};\nconst manager = new OutputManager({ loader: shinyRequireLoader });\n/******************************************************************************\n* Define the Shiny binding\n******************************************************************************/\n// Ideally we'd extend Shiny's HTMLOutputBinding, but the implementation isn't exported\nclass IPyWidgetOutput extends Shiny.OutputBinding {\n    find(scope) {\n        return $(scope).find(\".shiny-ipywidget-output\");\n    }\n    onValueError(el, err) {\n        Shiny.unbindAll(el);\n        this.renderError(el, err);\n    }\n    async renderValue(el, data) {\n        // Allow for a None/null value to hide the widget (css inspired by htmlwidgets)\n        if (!data) {\n            el.style.visibility = \"hidden\";\n            return;\n        }\n        else {\n            el.style.visibility = \"hidden\";\n        }\n        // Only forward the potential to fill if `output_widget(fillable=True)`\n        // _and_ the widget instance wants to fill\n        const fill = data.fill && el.classList.contains(\"html-fill-container\");\n        if (fill)\n            el.classList.add(\"forward-fill-potential\");\n        // At this time point, we should've already handled an 'open' message, and so\n        // the model should be ready to use\n        const model = await manager.get_model(data.model_id);\n        if (!model) {\n            throw new Error(`No model found for id ${data.model_id}`);\n        }\n        const view = await manager.create_view(model, {});\n        await manager.display_view(view, { el: el });\n        // Don't allow more than one .lmWidget container, which can happen\n        // when the view is displayed more than once\n        // N.B. It's probably better to get view(s) from m.views and .remove() them,\n        // but empirically, this seems to work better\n        while (el.childNodes.length > 1) {\n            el.removeChild(el.childNodes[0]);\n        }\n        // The ipywidgets container (.lmWidget)\n        const lmWidget = el.children[0];\n        if (fill) {\n            this._onImplementation(lmWidget, () => this._doAddFillClasses(lmWidget));\n        }\n        this._onImplementation(lmWidget, () => {\n            if (data.widget_pkg !== \"plotly\") {\n                this._doResize();\n                el.style.visibility = \"inherit\";\n                return;\n            }\n            const plotlyGraphDiv = this._findPlotlyGraphDiv(lmWidget);\n            if (!plotlyGraphDiv) {\n                this._doResize();\n                el.style.visibility = \"inherit\";\n                return;\n            }\n            // Plotly FigureWidget may first render at its internal 360px fallback,\n            // then resize after paint. Keep it hidden until a direct Plotly resize\n            // completes so the first visible paint is already settled.\n            void this._settlePlotly(plotlyGraphDiv).then(() => {\n                el.style.visibility = \"inherit\";\n            });\n        });\n    }\n    async _settlePlotly(plotEl) {\n        var _a;\n        const plotly = window.Plotly;\n        await this._waitForPlotlyWidgetRender(plotEl);\n        await this._waitForPlotlyAfterPlot(plotEl);\n        if ((_a = plotly === null || plotly === void 0 ? void 0 : plotly.Plots) === null || _a === void 0 ? void 0 : _a.resize) {\n            const afterResize = this._waitForPlotlyAfterPlot(plotEl);\n            await plotly.Plots.resize(plotEl);\n            await afterResize;\n        }\n        else {\n            this._doResize();\n        }\n        this._doResize();\n    }\n    _onImplementation(lmWidget, callback) {\n        if (this._hasImplementation(lmWidget)) {\n            callback();\n            return;\n        }\n        // Some widget implementation (e.g., ipyleaflet, pydeck) won't actually\n        // have rendered to the DOM at this point, so wait until they do\n        const mo = new MutationObserver((_mutations) => {\n            if (this._hasImplementation(lmWidget)) {\n                mo.disconnect();\n                callback();\n            }\n        });\n        mo.observe(lmWidget, { childList: true });\n    }\n    _waitForPlotlyWidgetRender(plotEl) {\n        if (plotEl._fullLayout) {\n            return Promise.resolve();\n        }\n        return new Promise((resolve) => {\n            const onRender = (evt) => {\n                var _a;\n                const target = (_a = evt.detail) === null || _a === void 0 ? void 0 : _a.element;\n                if (target !== plotEl)\n                    return;\n                cleanup();\n                resolve();\n            };\n            const cleanup = () => {\n                document.removeEventListener(\"plotlywidget-after-render\", onRender);\n                window.clearTimeout(timeoutId);\n            };\n            const timeoutId = window.setTimeout(() => {\n                cleanup();\n                resolve();\n            }, 1000);\n            document.addEventListener(\"plotlywidget-after-render\", onRender);\n        });\n    }\n    _waitForPlotlyAfterPlot(plotEl) {\n        return new Promise((resolve) => {\n            const handler = () => {\n                cleanup();\n                resolve();\n            };\n            const cleanup = () => {\n                if (hasMethod(plotEl, \"removeListener\")) {\n                    plotEl.removeListener(\"plotly_afterplot\", handler);\n                }\n                window.clearTimeout(timeoutId);\n            };\n            const timeoutId = window.setTimeout(() => {\n                cleanup();\n                resolve();\n            }, 1000);\n            if (hasMethod(plotEl, \"once\")) {\n                plotEl.once(\"plotly_afterplot\", handler);\n                return;\n            }\n            if (hasMethod(plotEl, \"on\")) {\n                plotEl.on(\"plotly_afterplot\", handler);\n                return;\n            }\n            cleanup();\n            resolve();\n        });\n    }\n    // In most cases, we can get widgets to fill through Python/CSS, but some widgets\n    // (e.g., quak) don't have a Python API and use shadow DOM, which can only access\n    // from JS\n    _doAddFillClasses(lmWidget) {\n        var _a;\n        const impl = lmWidget.children[0];\n        const isQuakWidget = impl && !!((_a = impl.shadowRoot) === null || _a === void 0 ? void 0 : _a.querySelector(\".quak\"));\n        if (isQuakWidget) {\n            impl.classList.add(\"html-fill-container\", \"html-fill-item\");\n            const quakWidget = impl.shadowRoot.querySelector(\".quak\");\n            quakWidget.style.maxHeight = \"unset\";\n        }\n    }\n    _doResize() {\n        // Trigger resize event to force layout (setTimeout() is needed for altair)\n        // TODO: debounce this call?\n        setTimeout(() => {\n            window.dispatchEvent(new Event('resize'));\n        }, 0);\n    }\n    _findPlotlyGraphDiv(root) {\n        const plotlyEl = root.matches(\".js-plotly-plot\")\n            ? root\n            : root.querySelector(\".js-plotly-plot\");\n        return plotlyEl instanceof HTMLElement ? plotlyEl : null;\n    }\n    _hasImplementation(lmWidget) {\n        var _a;\n        const impl = lmWidget.children[0];\n        return impl && (impl.children.length > 0 || ((_a = impl.shadowRoot) === null || _a === void 0 ? void 0 : _a.children.length) > 0);\n    }\n}\nShiny.outputBindings.register(new IPyWidgetOutput(), \"shiny.IPyWidgetOutput\");\n// Due to the way HTMLManager (and widget implementations) get loaded (via\n// require.js), the binding registration above can happen _after_ Shiny has\n// already bound the DOM, especially in the dynamic UI case (i.e., output_binding()'s\n// dependencies don't come in until after initial page load). And, in the dynamic UI\n// case, UI is rendered asychronously via Shiny.shinyapp.taskQueue, so if it exists,\n// we probably need to re-bind the DOM after the taskQueue is done.\nconst taskQueue = (_a = Shiny === null || Shiny === void 0 ? void 0 : Shiny.shinyapp) === null || _a === void 0 ? void 0 : _a.taskQueue;\nif (taskQueue) {\n    taskQueue.enqueue(() => Shiny.bindAll(document.body));\n}\n/******************************************************************************\n* Handle messages from the server-side Widget\n******************************************************************************/\n// Initialize the comm and model when a new widget is created\n// This is basically our version of https://github.com/jupyterlab/jupyterlab/blob/d33de15/packages/services/src/kernel/default.ts#L1144-L1176\nShiny.addCustomMessageHandler(\"shinywidgets_comm_open\", (msg_txt) => {\n    setBaseURL();\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_2__.jsonParse)(msg_txt);\n    Shiny.renderDependencies(msg.content.html_deps);\n    const comm = new _comm__WEBPACK_IMPORTED_MODULE_1__.ShinyComm(msg.content.comm_id);\n    manager.handle_comm_open(comm, msg);\n});\n// Handle any mutation of the model (e.g., add a marker to a map, without a full redraw)\n// Basically out version of https://github.com/jupyterlab/jupyterlab/blob/d33de15/packages/services/src/kernel/default.ts#L1200-L1215\nShiny.addCustomMessageHandler(\"shinywidgets_comm_msg\", async (msg_txt) => {\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_2__.jsonParse)(msg_txt);\n    const id = msg.content.comm_id;\n    const model = manager.get_model(id);\n    if (!model) {\n        console.error(`Couldn't handle message for model ${id} because it doesn't exist.`);\n        return;\n    }\n    try {\n        const m = await model;\n        // @ts-ignore for some reason IClassicComm doesn't have this method, but we do\n        m.comm.handle_msg(msg);\n    }\n    catch (err) {\n        console.error(\"Error handling message:\", err);\n    }\n});\n// Handle the closing of a widget/comm/model\nShiny.addCustomMessageHandler(\"shinywidgets_comm_close\", async (msg_txt) => {\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_2__.jsonParse)(msg_txt);\n    const id = msg.content.comm_id;\n    const model = manager.get_model(id);\n    if (!model) {\n        console.error(`Couldn't close model ${id} because it doesn't exist.`);\n        return;\n    }\n    try {\n        const m = await model;\n        // Some widget views need explicit teardown before model.close() removes them.\n        if (m.views) {\n            await Promise.all(Object.values(m.views).map(async (viewPromise) => {\n                try {\n                    const v = await viewPromise;\n                    // Plotly-backed views can leave DOM state and listeners behind unless\n                    // destroy() runs before the view is removed.\n                    if (hasMethod(v, 'destroy')) {\n                        v.destroy();\n                        // Clearing the back-reference prevents later teardown from touching a\n                        // model that is already being closed.\n                        delete v.model;\n                        v.remove();\n                    }\n                }\n                catch (err) {\n                    console.error(\"Error cleaning up view:\", err);\n                }\n            }));\n        }\n        // View removal updates _view_count. Mark the comm as inactive first so that\n        // save_changes() does not try to send those updates after the comm is gone.\n        m.comm_live = false;\n        // Close model after all views are cleaned up.\n        try {\n            await m.close();\n        }\n        catch (closeErr) {\n            if (!isIgnorableTeardownError(closeErr)) {\n                console.error(\"Unexpected error while closing model:\", closeErr);\n            }\n        }\n        // HTMLManager releases the model from its registry on comm:close.\n        try {\n            m.trigger(\"comm:close\");\n        }\n        catch (triggerErr) {\n            if (!isIgnorableTeardownError(triggerErr)) {\n                console.error(\"Unexpected error while triggering comm:close:\", triggerErr);\n            }\n        }\n    }\n    catch (err) {\n        console.error(\"Error during model cleanup:\", err);\n    }\n});\n$(document).on(\"shiny:disconnected\", () => {\n    manager.clear_state();\n});\n// When in filling layout, some widgets (specifically, altair) incorrectly think their\n// height is 0 after it's shown, hidden, then shown again. As a workaround, trigger a\n// resize event when a tab is shown.\n// TODO: This covers the 95% use case, but it's definitely not an ideal way to handle\n// this situation. A more robust solution would use IntersectionObserver to detect when\n// the widget becomes visible. Or better yet, we'd get altair to handle this situation\n// better.\n// https://github.com/posit-dev/py-shinywidgets/issues/172\ndocument.addEventListener('shown.bs.tab', event => {\n    window.dispatchEvent(new Event('resize'));\n});\n// Our version of https://github.com/jupyter-widgets/widget-cookiecutter/blob/9694718/%7B%7Bcookiecutter.github_project_name%7D%7D/js/lib/extension.js#L8\nfunction setBaseURL(x = '') {\n    const base_url = document.querySelector('body').getAttribute('data-base-url');\n    if (!base_url) {\n        document.querySelector('body').setAttribute('data-base-url', x);\n    }\n}\n// TypeGuard to safely check if an object has a method\nfunction hasMethod(obj, methodName) {\n    return typeof obj[methodName] === 'function';\n}\nfunction isIgnorableTeardownError(err) {\n    const msg = errorMessage(err).toLowerCase();\n    return (msg.includes(\"widget is not attached\") ||\n        msg.includes(\"no comm channel defined\"));\n}\nfunction errorMessage(err) {\n    if (err instanceof Error) {\n        return err.message;\n    }\n    return String(err);\n}\n\n\n//# sourceURL=webpack://@jupyter-widgets/shiny-embed-manager/./src/output.ts?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @jupyter-widgets/html-manager */ \"@jupyter-widgets/html-manager\");\n/* harmony import */ var _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _comm__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./comm */ \"./src/comm.ts\");\n/* harmony import */ var _plotly__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./plotly */ \"./src/plotly.ts\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./utils */ \"./src/utils.ts\");\nvar _a;\n\n\n\n\n/******************************************************************************\n * Define a custom HTMLManager for use with Shiny\n ******************************************************************************/\nclass OutputManager extends _jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__.HTMLManager {\n    // In a soon-to-be-released version of @jupyter-widgets/html-manager,\n    // display_view()'s first \"dummy\" argument will be removed... this shim simply\n    // makes it so that our manager can work with either version\n    // https://github.com/jupyter-widgets/ipywidgets/commit/159bbe4#diff-45c126b24c3c43d2cee5313364805c025e911c4721d45ff8a68356a215bfb6c8R42-R43\n    async display_view(view, options) {\n        const n_args = super.display_view.length;\n        if (n_args === 3) {\n            return super.display_view({}, view, options);\n        }\n        else {\n            // @ts-ignore\n            return super.display_view(view, options);\n        }\n    }\n}\n// Define our own custom module loader for Shiny\nconst shinyRequireLoader = async function (moduleName, moduleVersion) {\n    // shiny provides a shim of require.js which allows <script>s with anonymous\n    // define()s to be loaded without error. When an anonymous define() occurs,\n    // the shim uses the data-requiremodule attribute (set by require.js) on the script\n    // to determine the module name.\n    // https://github.com/posit-dev/py-shiny/blob/230940c/scripts/define-shims.js#L10-L16\n    // In the context of shinywidgets, when a widget gets rendered, it should\n    // come with another <script> tag that does `require.config({paths: {...}})`\n    // which maps the module name to a URL of the widget's JS file.\n    const oldAmd = window.define.amd;\n    // This is probably not necessary, but just in case -- especially now in a\n    // anywidget/ES6 world, we probably don't want to load AMD modules\n    // (plotly is one example of a widget that will fail to load if AMD is enabled)\n    window.define.amd = false;\n    // Store jQuery global since loading we load a module, it may overwrite it\n    // (qgrid is one good example)\n    const old$ = window.$;\n    const oldJQ = window.jQuery;\n    if (moduleName === 'qgrid') {\n        // qgrid wants to use base/js/dialog (if it's available) for full-screen tables\n        // https://github.com/quantopian/qgrid/blob/877b420/js/src/qgrid.widget.js#L11-L16\n        // Maybe that's worth supporting someday, but for now, we define it to be nothing\n        // to avoid require('qgrid') from producing an error\n        window.define(\"base/js/dialog\", [], function () { return null; });\n    }\n    return (0,_jupyter_widgets_html_manager__WEBPACK_IMPORTED_MODULE_0__.requireLoader)(moduleName, moduleVersion).finally(() => {\n        window.define.amd = oldAmd;\n        window.$ = old$;\n        window.jQuery = oldJQ;\n    });\n};\nconst manager = new OutputManager({ loader: shinyRequireLoader });\n/******************************************************************************\n* Define the Shiny binding\n******************************************************************************/\n// Ideally we'd extend Shiny's HTMLOutputBinding, but the implementation isn't exported\nclass IPyWidgetOutput extends Shiny.OutputBinding {\n    find(scope) {\n        return $(scope).find(\".shiny-ipywidget-output\");\n    }\n    onValueError(el, err) {\n        Shiny.unbindAll(el);\n        this.renderError(el, err);\n    }\n    async renderValue(el, data) {\n        const hiddenVisibility = \"hidden\";\n        const revealVisibility = \"inherit\";\n        // Allow for a None/null value to hide the widget (css inspired by htmlwidgets)\n        if (!data) {\n            el.style.visibility = hiddenVisibility;\n            return;\n        }\n        // Treat widget outputs as hidden by default and only reveal them once\n        // their initial display conditions are satisfied.\n        el.style.visibility = hiddenVisibility;\n        // Only forward the potential to fill if `output_widget(fillable=True)`\n        // _and_ the widget instance wants to fill\n        const fill = data.fill && el.classList.contains(\"html-fill-container\");\n        if (fill)\n            el.classList.add(\"forward-fill-potential\");\n        // At this time point, we should've already handled an 'open' message, and so\n        // the model should be ready to use\n        const model = await manager.get_model(data.model_id);\n        if (!model) {\n            throw new Error(`No model found for id ${data.model_id}`);\n        }\n        const view = await manager.create_view(model, {});\n        await manager.display_view(view, { el: el });\n        // Don't allow more than one .lmWidget container, which can happen\n        // when the view is displayed more than once\n        // N.B. It's probably better to get view(s) from m.views and .remove() them,\n        // but empirically, this seems to work better\n        while (el.childNodes.length > 1) {\n            el.removeChild(el.childNodes[0]);\n        }\n        // The ipywidgets container (.lmWidget)\n        const lmWidget = el.children[0];\n        if (fill) {\n            this._onImplementation(lmWidget, () => this._doAddFillClasses(lmWidget));\n        }\n        this._onImplementation(lmWidget, () => {\n            if (data.widget_pkg !== \"plotly\") {\n                this._doResize();\n                el.style.visibility = revealVisibility;\n                return;\n            }\n            const plotlyGraphDiv = (0,_plotly__WEBPACK_IMPORTED_MODULE_2__.findPlotlyGraphDiv)(lmWidget);\n            if (!plotlyGraphDiv) {\n                this._doResize();\n                el.style.visibility = revealVisibility;\n                return;\n            }\n            // Plotly FigureWidget may first render at its internal 360px fallback,\n            // then resize after paint. Keep it hidden until a direct Plotly resize\n            // completes so the first visible paint is already settled.\n            void this._settlePlotly(plotlyGraphDiv).then(() => {\n                el.style.visibility = revealVisibility;\n            });\n        });\n    }\n    async _settlePlotly(plotEl) {\n        await (0,_plotly__WEBPACK_IMPORTED_MODULE_2__.settlePlotlyOutput)(plotEl, () => this._doResize());\n    }\n    _onImplementation(lmWidget, callback) {\n        if (this._hasImplementation(lmWidget)) {\n            callback();\n            return;\n        }\n        // Some widget implementation (e.g., ipyleaflet, pydeck) won't actually\n        // have rendered to the DOM at this point, so wait until they do\n        const mo = new MutationObserver((_mutations) => {\n            if (this._hasImplementation(lmWidget)) {\n                mo.disconnect();\n                callback();\n            }\n        });\n        mo.observe(lmWidget, { childList: true });\n    }\n    // In most cases, we can get widgets to fill through Python/CSS, but some widgets\n    // (e.g., quak) don't have a Python API and use shadow DOM, which can only access\n    // from JS\n    _doAddFillClasses(lmWidget) {\n        var _a;\n        const impl = lmWidget.children[0];\n        const isQuakWidget = impl && !!((_a = impl.shadowRoot) === null || _a === void 0 ? void 0 : _a.querySelector(\".quak\"));\n        if (isQuakWidget) {\n            impl.classList.add(\"html-fill-container\", \"html-fill-item\");\n            const quakWidget = impl.shadowRoot.querySelector(\".quak\");\n            quakWidget.style.maxHeight = \"unset\";\n        }\n    }\n    _doResize() {\n        // Trigger resize event to force layout (setTimeout() is needed for altair)\n        // TODO: debounce this call?\n        setTimeout(() => {\n            window.dispatchEvent(new Event('resize'));\n        }, 0);\n    }\n    _hasImplementation(lmWidget) {\n        var _a;\n        const impl = lmWidget.children[0];\n        return impl && (impl.children.length > 0 || ((_a = impl.shadowRoot) === null || _a === void 0 ? void 0 : _a.children.length) > 0);\n    }\n}\nShiny.outputBindings.register(new IPyWidgetOutput(), \"shiny.IPyWidgetOutput\");\n// Due to the way HTMLManager (and widget implementations) get loaded (via\n// require.js), the binding registration above can happen _after_ Shiny has\n// already bound the DOM, especially in the dynamic UI case (i.e., output_binding()'s\n// dependencies don't come in until after initial page load). And, in the dynamic UI\n// case, UI is rendered asychronously via Shiny.shinyapp.taskQueue, so if it exists,\n// we probably need to re-bind the DOM after the taskQueue is done.\nconst taskQueue = (_a = Shiny === null || Shiny === void 0 ? void 0 : Shiny.shinyapp) === null || _a === void 0 ? void 0 : _a.taskQueue;\nif (taskQueue) {\n    taskQueue.enqueue(() => Shiny.bindAll(document.body));\n}\n/******************************************************************************\n* Handle messages from the server-side Widget\n******************************************************************************/\n// Initialize the comm and model when a new widget is created\n// This is basically our version of https://github.com/jupyterlab/jupyterlab/blob/d33de15/packages/services/src/kernel/default.ts#L1144-L1176\nShiny.addCustomMessageHandler(\"shinywidgets_comm_open\", (msg_txt) => {\n    setBaseURL();\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_3__.jsonParse)(msg_txt);\n    Shiny.renderDependencies(msg.content.html_deps);\n    const comm = new _comm__WEBPACK_IMPORTED_MODULE_1__.ShinyComm(msg.content.comm_id);\n    manager.handle_comm_open(comm, msg);\n});\n// Handle any mutation of the model (e.g., add a marker to a map, without a full redraw)\n// Basically out version of https://github.com/jupyterlab/jupyterlab/blob/d33de15/packages/services/src/kernel/default.ts#L1200-L1215\nShiny.addCustomMessageHandler(\"shinywidgets_comm_msg\", async (msg_txt) => {\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_3__.jsonParse)(msg_txt);\n    const id = msg.content.comm_id;\n    const model = manager.get_model(id);\n    if (!model) {\n        console.error(`Couldn't handle message for model ${id} because it doesn't exist.`);\n        return;\n    }\n    try {\n        const m = await model;\n        // @ts-ignore for some reason IClassicComm doesn't have this method, but we do\n        m.comm.handle_msg(msg);\n    }\n    catch (err) {\n        console.error(\"Error handling message:\", err);\n    }\n});\n// Handle the closing of a widget/comm/model\nShiny.addCustomMessageHandler(\"shinywidgets_comm_close\", async (msg_txt) => {\n    const msg = (0,_utils__WEBPACK_IMPORTED_MODULE_3__.jsonParse)(msg_txt);\n    const id = msg.content.comm_id;\n    const model = manager.get_model(id);\n    if (!model) {\n        console.error(`Couldn't close model ${id} because it doesn't exist.`);\n        return;\n    }\n    try {\n        const m = await model;\n        // Some widget views need explicit teardown before model.close() removes them.\n        if (m.views) {\n            await Promise.all(Object.values(m.views).map(async (viewPromise) => {\n                try {\n                    const v = await viewPromise;\n                    // Plotly-backed views can leave DOM state and listeners behind unless\n                    // destroy() runs before the view is removed.\n                    if (hasMethod(v, 'destroy')) {\n                        v.destroy();\n                        // Clearing the back-reference prevents later teardown from touching a\n                        // model that is already being closed.\n                        delete v.model;\n                        v.remove();\n                    }\n                }\n                catch (err) {\n                    console.error(\"Error cleaning up view:\", err);\n                }\n            }));\n        }\n        // View removal updates _view_count. Mark the comm as inactive first so that\n        // save_changes() does not try to send those updates after the comm is gone.\n        m.comm_live = false;\n        // Close model after all views are cleaned up.\n        try {\n            await m.close();\n        }\n        catch (closeErr) {\n            if (!isIgnorableTeardownError(closeErr)) {\n                console.error(\"Unexpected error while closing model:\", closeErr);\n            }\n        }\n        // HTMLManager releases the model from its registry on comm:close.\n        try {\n            m.trigger(\"comm:close\");\n        }\n        catch (triggerErr) {\n            if (!isIgnorableTeardownError(triggerErr)) {\n                console.error(\"Unexpected error while triggering comm:close:\", triggerErr);\n            }\n        }\n    }\n    catch (err) {\n        console.error(\"Error during model cleanup:\", err);\n    }\n});\n$(document).on(\"shiny:disconnected\", () => {\n    manager.clear_state();\n});\n// When in filling layout, some widgets (specifically, altair) incorrectly think their\n// height is 0 after it's shown, hidden, then shown again. As a workaround, trigger a\n// resize event when a tab is shown.\n// TODO: This covers the 95% use case, but it's definitely not an ideal way to handle\n// this situation. A more robust solution would use IntersectionObserver to detect when\n// the widget becomes visible. Or better yet, we'd get altair to handle this situation\n// better.\n// https://github.com/posit-dev/py-shinywidgets/issues/172\ndocument.addEventListener('shown.bs.tab', event => {\n    window.dispatchEvent(new Event('resize'));\n});\n// Our version of https://github.com/jupyter-widgets/widget-cookiecutter/blob/9694718/%7B%7Bcookiecutter.github_project_name%7D%7D/js/lib/extension.js#L8\nfunction setBaseURL(x = '') {\n    const base_url = document.querySelector('body').getAttribute('data-base-url');\n    if (!base_url) {\n        document.querySelector('body').setAttribute('data-base-url', x);\n    }\n}\n// TypeGuard to safely check if an object has a method\nfunction hasMethod(obj, methodName) {\n    return typeof obj[methodName] === 'function';\n}\nfunction isIgnorableTeardownError(err) {\n    const msg = errorMessage(err).toLowerCase();\n    return (msg.includes(\"widget is not attached\") ||\n        msg.includes(\"no comm channel defined\"));\n}\nfunction errorMessage(err) {\n    if (err instanceof Error) {\n        return err.message;\n    }\n    return String(err);\n}\n\n\n//# sourceURL=webpack://@jupyter-widgets/shiny-embed-manager/./src/output.ts?");
+
+/***/ }),
+
+/***/ "./src/plotly.ts":
+/*!***********************!*\
+  !*** ./src/plotly.ts ***!
+  \***********************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"findPlotlyGraphDiv\": () => (/* binding */ findPlotlyGraphDiv),\n/* harmony export */   \"settlePlotlyOutput\": () => (/* binding */ settlePlotlyOutput)\n/* harmony export */ });\nfunction findPlotlyGraphDiv(root) {\n    const plotlyEl = root.matches(\".js-plotly-plot\")\n        ? root\n        : root.querySelector(\".js-plotly-plot\");\n    return plotlyEl instanceof HTMLElement ? plotlyEl : null;\n}\nasync function settlePlotlyOutput(plotEl, dispatchResize) {\n    var _a;\n    const plotly = window.Plotly;\n    await waitForPlotlyWidgetRender(plotEl);\n    await waitForPlotlyAfterPlot(plotEl);\n    if ((_a = plotly === null || plotly === void 0 ? void 0 : plotly.Plots) === null || _a === void 0 ? void 0 : _a.resize) {\n        const afterResize = waitForPlotlyAfterPlot(plotEl);\n        await plotly.Plots.resize(plotEl);\n        await afterResize;\n    }\n    else {\n        dispatchResize();\n    }\n    dispatchResize();\n}\nfunction waitForPlotlyWidgetRender(plotEl) {\n    if (plotEl._fullLayout) {\n        return Promise.resolve();\n    }\n    return new Promise((resolve) => {\n        const onRender = (evt) => {\n            var _a;\n            const target = (_a = evt.detail) === null || _a === void 0 ? void 0 : _a.element;\n            if (target !== plotEl)\n                return;\n            cleanup();\n            resolve();\n        };\n        const cleanup = () => {\n            document.removeEventListener(\"plotlywidget-after-render\", onRender);\n            window.clearTimeout(timeoutId);\n        };\n        const timeoutId = window.setTimeout(() => {\n            cleanup();\n            resolve();\n        }, 1000);\n        document.addEventListener(\"plotlywidget-after-render\", onRender);\n    });\n}\nfunction waitForPlotlyAfterPlot(plotEl) {\n    return new Promise((resolve) => {\n        const handler = () => {\n            cleanup();\n            resolve();\n        };\n        const cleanup = () => {\n            if (hasMethod(plotEl, \"removeListener\")) {\n                plotEl.removeListener(\"plotly_afterplot\", handler);\n            }\n            window.clearTimeout(timeoutId);\n        };\n        const timeoutId = window.setTimeout(() => {\n            cleanup();\n            resolve();\n        }, 1000);\n        if (hasMethod(plotEl, \"once\")) {\n            plotEl.once(\"plotly_afterplot\", handler);\n            return;\n        }\n        if (hasMethod(plotEl, \"on\")) {\n            plotEl.on(\"plotly_afterplot\", handler);\n            return;\n        }\n        cleanup();\n        resolve();\n    });\n}\nfunction hasMethod(x, key) {\n    return !!x && typeof x[key] === \"function\";\n}\n\n\n//# sourceURL=webpack://@jupyter-widgets/shiny-embed-manager/./src/plotly.ts?");
 
 /***/ }),
